requirements.txt
flask
numpy
pillow
fastai
gunicorn==19.9.0